Load initial data on instance start
In OS4, the database setup is handled by the datastore and manage
services.  The instance must be started at which point the datastore
creates the basic schema.  Afterwards, the openslides tool (manage
service) must be run explicitly to load the so-called initial data into
the database.

The instance creation workflow chosen here is to run `openslides
initial-data` with every start of an instance.  This should be a no-op
in case the command has been run before.