ls: List running image tags
With OpenSlides 3, osinstancectl would query the backend for its
version.  This version was technically independent from image tags and
a valuable data point.  Furthermore, it acted as a rather reliable
indicator for an instance's health because most of the time when there
was a problem with an OS3 instance, the version retrieval would also
fail.

OpenSlides 4 no longer keeps track of its own version.  In order to
display something useful in ls's version field, osinstancectl will
retrieve the actually running versions of all the instances' images from
Docker.  This will avoid getting incorrect results from merely edited
but not deployed configuration files (the version information in ls -l
can get fooled that way).  If not all images carry the same tag,
multiple tags are listed.

While this version information is hopefully going to be somewhat useful,
it does not replace the health check aspect of the OS3 procedure.  As
a reminder, ls will display running instances with the "unknown" status
(yellow) as a best case scenario.