osinstancectl: Check instance health during ls
The backend service's health status can now be queried by HTTP.  The
backend's health is used as an *indicator* for the instance's overall
health, even though the metric is too simple for that.  It is entirely
possible for osinstancectl to list an instance as healthy that obviously
isn't.  Yet, the situation is an improvement over no indicator at all
because it at least shows that the instance was able to fully start up
at some point and that the backend is still up.

This patch reintroduces the green/"OK" status to ls.