osinstancectl: Re-add function to create user account
This patch reintroduces functionality that already existed for
OpenSlides 3 instances which is adding a second admin user account to
new instances.  It had to be removed due to the changes required for
OpenSlides 4.

Most notably, the 'openslides' management tool is now used to create
user accounts.  Previously, this was handled by the backend's startup
routine/entrypoint.  Unlike with OpenSlides 3, new instances must be
started to finish the setup.

Please note that at this point the user accounts created by 'openslides'
are inactive and lack sufficient access rights to be useful.  Once the
management tool is fixed, additional fields will likely need to be added
to user.yml.