osinstancectl: Extended version info format for ls
In its compact form, `ls` already includes a list of tags used by an
instance.  This patch appends the number of registries and tags.

This information is important because osinstancectl's update function
only updates the default tag.  If there is more than one tag and/or one
registry in use, it means that osinstancectl cannot fully update the
instance automatically.