osinstancectl: Run setup commands on instance start
By checking the new 'openslides initial-data' exit code and adding a new
method of managing osinstancectl's own user setup YAML files it is now
safe to run the 'openslides' setup commands on each start.  That means
it is no longer necessary to start an instance immediately after
creating it.  Instead, the final setup steps will be executed during the
first time the instance is started.