Add bulk update script
This version is based on the OpenSlides 3 bulk update script with
various improvements.

It does not yet support osinstancectl's --management-tool option because
it remains to be seen if it is really necessary.

This patch also removes a space from osinstancectl's version string
listing to simplify field splitting by the bulk update script.