osinstancectl: Remove --www option
In commit ea92ccf, the ALLOWED_HOSTS functionality was added for
OpenSlides 4 but it is ignoring the --www option.  Instead of fixing
that bug, this may be an opportune time to remove the option instead.
It has not been used in a while and probably not been tested very well
with OpenSlides 4.