osinstancectl: Dynamically generate ls tree structure
By generating the tree strcuture dynamically, osinstancectl can be more
easily adapted for complex information hierarchies which are expected
for OpenSlides 4.

As an immediate improvement, the user secrets listing has been separated
into its own branch.  The JSON output format has not been altered,
however, so dependent tools should not be affected by this change.

On the downside, the listing has become (even) slower with this change.
Surely, there is lots of room for optimizations.