osinstancectl: Improve ls --stats
For ls --stats, the data collected for the regular and the JSON output
formats differs.  This patch adds additional conditionals to avoid
unnecessary queries.

Additionally, a lot of the output formatting for --stats was prepared
early during the data collection phase.  This formatting has been moved
to the tree formatting stage.