bulk-update: Run jobs in parallel
Parallel execution is handled by GNU parallel.  The --jobs option
parameter is passed through to parallel, so see the parallel(1)
documentation for advanced usage such as percentages.

Without additional options, osinstancectl's detailed output is printed
on the screen once completed as well as written to a directory.  With
the options --tmux or --tmuxpanes, however, the output is only displayed
in a tmux session instead.

In any case, osinstancectl writes its own logs of the update process in
the individual instance directories.

This patch also removes the --at functionality.  Considering the
long-running nature of OpenSlides 4 data migrations, it becomes almost
impossible to schedule them to finish, e.g., during the night.
Furthermore, it is now safe to begin migrations at any time without
disrupting operations.  Eventually, we are going to need more control
over when the finalize command is run, however.