Include version string from image
OpenSlides 4 includes its own version string again, similar to
OpenSlides 3.  This string can be retrieved with just curl and without
access to Docker.  It can be used as a fallback for users who lack
docker access to view the actual image versions.

For now, the image/tag-derived version string remains the default
because, arguably, it contains more useful information but it might make
sense to reverse that.