bulk-update: Always set management tool version
Unless a management tool version is configured for the individual
instances with `osinstancectl --management-tool`, osinstancectl will
always use the latest openslides management tool version available for
start, stop, update etc.  In an environment running multiple versions,
this can lead to incompatibilities.  It is safer to pin down the correct
version for each instance.  This patch allows the bulk update script to
pass this option to osinstancectl during updates.