Enforce pinning of the management tool version
This patch reverses osinstancectl's default behavior regarding the
management tool version associated with an instance.  Before, unless the
option --management-tool was given, osinstancectl would configure the
instance to always use the newest version available at any given time.

In order to increase its use, this version of osinstancectl records and
pins the precise management tool version at the time an instance is
created.  It follows that for updates, the version must be kept in sync
with the image versions, so the option, like --tag, is now mandatory.

The special string '-' is now a shortcut to indicate that the version is
to remain unchanged, making simple updates barely any more complicated.
The special string '*' carries the previous meaning of '-', i.e.,
"always use the newest available version".