Add instance locking functionality
Instance locks prohibit some or all actions to be executed for a given
instance.

The lock status contains a message which is displayed as part of error
messages as well as `ls --long`.  The mechanism is meant to help enforce
communications that may have otherwise happend through the --metadata
feature or another channel, e.g., "do not update this instance".

The --json output is incomplete as it does not include the individual
action locks but merely boolean fields.  This is good enough, however,
to enable the bulk update script to detect locked instances.  It should
be relatively straight forward to add a complete list of locks.