Add setup command
This commit mainly adds a `setup` command for osinstancectl that,
hopefully, makes it easier to get started.

The UI leaves lots of room for improvement; however, this is a very
simple implementation that should be easy to adapt as we learn which
checks are helpful and what is missing.

To make the setup command's output less verbose, this patch also adds
a --quiet option to openslides-bin-installer.