Offer to create minimal config.yml template if necessary
osinstancectl requires configurations that differ from the OpenSlides
management tool's defaults.  For example, the Postgres service and HTTPS
features must be disabled.  For this reason, a custom management tool
configuration file must be present.  osinstancectl will refuse to
continue without a configuration file but offer to a create minimal one.

osinstancectl could alternatively quietly apply the required custom
settings without insisting on a static configuration file.  There are
probably good arguments in favor of it; however, the method chosen here
has the advantage of being fully transparent.  Additionally, the
generated configuration file can be used with the management tool
directly which users may expect to be able to do.