Make --management-tool optional for updates
For OpenSlides 4 instances, the 'openslides' management tool must be
compatible with its accompanying 'manage' service, so during updates,
both the Docker services and the management tool must be updated in
tandem.

Most of the time, instances are updated to the latest version, making
--management-tool=latest the correct option.  While 'latest' was
initially intended as a default value, osinstancectl explicitiely
required the option as a precaution.  Day-to-day use has confirmed that
'latest' is indeed suitable as a default, however.